Make the NamedMatcher a 1st class citizen of combination matchers
